dataflow/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#
setuptools
apache-beam[gcp]==2.36.0
google-cloud-bigquery>=1.27.2
google-cloud-storage>=1.35.0
google-cloud-bigquery-storage==2.6.3
gunicorn==20.0.4
pandas>=1.1.3
pandas-gbq>=0.17.0
xgboost>=0.90
scikit-learn>=1.0.1
imbalanced-learn==0.9.0
pyarrow==3.0.0
